+++ e2e-tests/runner/examples/connectors/BurnerWalletConnector/BurnerWalletConnector.test.ts
@@ -44,7 +44,7 @@ test.describe('BurnerWalletConnector', async () => {
     const addressElement = await page.locator('css=#address');
 
     const address = await addressElement.getAttribute('data-address');
-    const amount: BN = bn(100_000_000);
+    const amount: BN = bn(2_000_000_000);
 
     if (address) {
       await seedWallet(