+++ packages/burner-wallet-connector/src/BurnerWalletConnector.ts
@@ -6,12 +6,14 @@ import {
   type ConnectorMetadata,
   FuelConnector,
   FuelConnectorEventTypes,
+  type HashableMessage,
   type JsonAbi,
   type Network,
   Provider,
   type SelectNetworkArguments,
   type StorageAbstract,
   type TransactionRequestLike,
+  type TransactionResponse,
   type Version,
   Wallet,
   type WalletUnlocked,
@@ -198,7 +200,10 @@ export class BurnerWalletConnector extends FuelConnector {
     return false;
   }
 
-  async signMessage(address: string, message: string): Promise<string> {
+  async signMessage(
+    address: string,
+    message: HashableMessage,
+  ): Promise<string> {
     if (!this.burnerWallet) {
       throw Error('Wallet not connected');
     }
@@ -215,7 +220,7 @@ export class BurnerWalletConnector extends FuelConnector {
   async sendTransaction(
     _address: string,
     transaction: TransactionRequestLike,
-  ): Promise<string> {
+  ): Promise<string | TransactionResponse> {
     if (!this.burnerWallet) {
       throw Error('Wallet not connected');
     }
@@ -224,10 +229,10 @@ export class BurnerWalletConnector extends FuelConnector {
       throw Error('Address not found for the connector');
     }
 
-    const transactionRequest =
+    const transactionResponse =
       await this.burnerWallet.sendTransaction(transaction);
 
-    return transactionRequest.id;
+    return transactionResponse;
   }
 
   async currentAccount(): Promise<string | null> {