+++ packages/fuel-wallet/src/FuelWalletConnector.ts
@@ -8,11 +8,15 @@ import {
   FuelConnectorEventType,
   FuelConnectorEventTypes,
   type FuelConnectorSendTxParams,
+  type HashableMessage,
   type Network,
   Provider,
   type SelectNetworkArguments,
+  type TransactionRequest,
   type TransactionRequestLike,
+  type TransactionResponse,
   type Version,
+  deserializeTransactionResponseJson,
   transactionRequestify,
 } from 'fuels';
 import type { JSONRPCRequest } from 'json-rpc-2.0';
@@ -149,6 +153,37 @@ export class FuelWalletConnector extends FuelConnector {
    * Connector methods
    * ============================================================
    */
+  private async prepareTransactionRequest(
+    transaction: TransactionRequestLike,
+    params?: FuelConnectorSendTxParams,
+  ) {
+    if (!transaction) {
+      throw new Error('Transaction is required');
+    }
+    let txRequest = transactionRequestify(transaction);
+
+    const {
+      onBeforeSend,
+      skipCustomFee,
+      transactionState,
+      transactionSummary,
+    } = params || {};
+
+    const providerToSend = params?.provider;
+
+    if (onBeforeSend) {
+      txRequest = await onBeforeSend(txRequest);
+    }
+
+    return {
+      txRequest,
+      providerToSend,
+      skipCustomFee,
+      transactionState,
+      transactionSummary,
+    };
+  }
+
   async ping(): Promise<boolean> {
     return this.client.timeout(800).request('ping', {});
   }
@@ -181,10 +216,21 @@ export class FuelWalletConnector extends FuelConnector {
     return Address.fromDynamicInput(account).toString();
   }
 
-  async signMessage(address: string, message: string): Promise<string> {
-    if (!message.trim()) {
+  async signMessage(
+    address: string,
+    message: HashableMessage,
+  ): Promise<string> {
+    if (typeof message === 'string' && !message.trim()) {
       throw new Error('Message is required');
     }
+    if (
+      typeof message === 'object' &&
+      message !== null &&
+      !message.personalSign
+    ) {
+      throw new Error('Message is required');
+    }
+
     return this.client.request('signMessage', {
       address,
       message,
@@ -195,34 +241,62 @@ export class FuelWalletConnector extends FuelConnector {
     address: string,
     transaction: TransactionRequestLike,
     params?: FuelConnectorSendTxParams,
-  ): Promise<string> {
-    if (!transaction) {
-      throw new Error('Transaction is required');
-    }
-    let txRequest = transactionRequestify(transaction);
-
+  ): Promise<string | TransactionResponse> {
     const {
-      onBeforeSend,
+      txRequest,
+      providerToSend,
       skipCustomFee,
-      provider,
       transactionState,
       transactionSummary,
-    } = params || {};
+    } = await this.prepareTransactionRequest(transaction, params);
 
-    if (onBeforeSend) {
-      txRequest = await onBeforeSend(txRequest);
-    }
-
-    // Transform transaction object to a transaction request
-
-    return this.client.request('sendTransaction', {
+    const resp = await this.client.request('sendTransaction', {
       address,
       transaction: JSON.stringify(txRequest),
-      provider,
+      provider: providerToSend,
       skipCustomFee,
       transactionState,
       transactionSummary,
+      returnTransactionResponse: true,
     });
+
+    if (typeof resp === 'object' && 'id' in resp && 'providerCache' in resp) {
+      return deserializeTransactionResponseJson(resp);
+    }
+
+    return resp?.id || resp;
+  }
+
+  async signTransaction(
+    address: string,
+    transaction: TransactionRequestLike,
+    params?: FuelConnectorSendTxParams,
+  ): Promise<TransactionRequest> {
+    const {
+      txRequest,
+      providerToSend,
+      skipCustomFee,
+      transactionState,
+      transactionSummary,
+    } = await this.prepareTransactionRequest(transaction, params);
+
+    const txRequestSerialized: string = await this.client.request(
+      'signTransaction',
+      {
+        address,
+        transaction: JSON.stringify(txRequest),
+        provider: providerToSend,
+        skipCustomFee,
+        transactionState,
+        transactionSummary,
+      },
+    );
+
+    const txRequestSigned = transactionRequestify(
+      JSON.parse(txRequestSerialized),
+    );
+
+    return txRequestSigned;
   }
 
   async assets(): Promise<Array<Asset>> {