+++ packages/react/src/ui/Connect/components/Connector/Connecting.tsx
@@ -74,6 +74,17 @@ export function Connecting({ className }: ConnectorProps) {
       }
 
       // If the connector is not native, let's check if we have already displayed the disclaimer
+      const supportsPredicateVersions =
+        connector &&
+        'getAvailablePredicateVersions' in connector &&
+        typeof (connector as { getAvailablePredicateVersions?: unknown })
+          .getAvailablePredicateVersions === 'function';
+
+      if (supportsPredicateVersions) {
+        setRoute(Routes.PredicateVersionSelector);
+        return;
+      }
+
       if (localStorage.getItem(PREDICATE_DISCLAIMER_KEY)) {
         cancel();
         return;
@@ -100,6 +111,14 @@ export function Connecting({ className }: ConnectorProps) {
       fuel.off(fuel.events.currentConnector, onCurrentConnectorChange);
     };
   }, [fuel]);
+
+  useEffect(() => {
+    if (error) {
+      if (error.message.includes('Failed to sign message')) {
+        setRoute(Routes.SignatureError);
+      }
+    }
+  }, [error, setRoute]);
 
   if (!connector) return null;