+++ packages/solana-connector/src/types.ts
@@ -1,6 +1,10 @@
 import type { PredicateConfig } from '@fuel-connectors/common';
 import type { ProviderType } from '@web3modal/solana/dist/types/src/utils/scaffold';
-import type { Provider as FuelProvider } from 'fuels';
+import type { Provider as FuelProvider, StorageAbstract } from 'fuels';
+
+export enum SolanaConnectorEvents {
+  ERROR = 'error',
+}
 
 export type SolanaConfig = {
   fuelProvider?: FuelProvider | Promise<FuelProvider>;
@@ -8,6 +12,7 @@ export type SolanaConfig = {
   predicateConfig?: PredicateConfig;
   solanaConfig?: ProviderType;
   chainId?: number;
+  storage?: StorageAbstract;
 };
 
 export interface GetAccounts {